Fix version script check for GNU BFD linker on FreeBSD
On FreeBSD, symbols 'environ' and ' __progname' are defined in crt1.o
and referenced by libc.so. This makes it different from platforms which
define 'environ' in libc.so. GNU BFD linker fails to link an executable
if these two symbols are not put into the global section of the version
script because a shared library should not reference a local symbol
defined in a different object file. GNU gold and LLVM LLD don't have
the problem.

This means the current check for --version-script fails on FreeBSD when
GNU BFD linker is used. The failure can be avoided by linking a shared
library instead of an executable, but meson doesn't support passing
additional arguments to the compiler in its 'has_link_argument' method.

Instead of adding these two unused symbols to the real version script,
use a smaller version script specifically made for the test.